feature[ats_VVMH0]: Integrate search engine into ORCS

+++ plugins/org.eclipse.osee.display.presenter.mocks/src/org/eclipse/osee/display/presenter/mocks/MockArtifact.java
@@ -98,7 +98,7 @@ public class MockArtifact implements ReadableArtifact {
    }
 
    @Override
-   public int getGammaId() {
+   public long getGammaId() {
       return 0;
    }
