+++ plugins/org.eclipse.osee.orcs.core/src/org/eclipse/osee/orcs/core/internal/artifact/Artifact.java
@@ -60,7 +60,7 @@ public class Artifact extends NamedIdentity<String> implements ReadableArtifact 
    }
 
    @Override
-   public int getGammaId() {
+   public long getGammaId() {
       return gammaId;
    }
